# backend/APIDocumentation.yaml
openapi: 3.0.0
info:
  title: Netro API
  description: |
    API for the Netro social platform. This API handles user authentication, profile management, 
    categories, email verification, and events.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## Response Format
    All responses follow a consistent format:
    - **Success responses** include `success: true` and a `message` field
    - **Error responses** include `success: false` and a descriptive `message` field
    - Validation errors include an additional `errors` array
    - Invalid category errors include an `invalidCategories` array
  version: 1.0.0
  contact:
    name: Netro API Support
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /profile/info:
    get:
      tags:
        - Profile
      summary: Get user profile
      description: Retrieve the authenticated user's profile information
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileSuccessResponse'
        '401':
          description: Unauthorized - missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                noToken:
                  value:
                    success: false
                    message: Access token is required
                invalidToken:
                  value:
                    success: false
                    message: Invalid or expired token
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: User not found
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: Failed to fetch user profile
  /profile/edit/info:
    put:
      tags:
        - Profile
      summary: Update user profile
      description: Update user profile information including username, location, bio, and categories
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userName:
                  type: string
                  maxLength: 100
                  description: User's display name
                  example: JohnDoe_Updated
                currentLocation:
                  type: string
                  maxLength: 100
                  description: User's current location
                  example: Los Angeles, CA
                bio:
                  type: string
                  maxLength: 100
                  description: Short biography
                  example: Full-stack developer and sports fan
                categories:
                  type: array
                  items:
                    type: string
                  description: Array of category names
                  example: ["basketball", "football", "programming"]
              minProperties: 1
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileUpdateSuccessResponse'
        '400':
          description: Validation failed or no fields provided
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ValidationError'
                  - $ref: '#/components/schemas/InvalidCategoriesError'
                  - $ref: '#/components/schemas/Error'
              examples:
                noFields:
                  value:
                    success: false
                    message: At least one field must be provided to update
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: User not found
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: Failed to update profile
  /profile/edit/password:
    put:
      tags:
        - Profile
      summary: Change password
      description: Change the authenticated user's password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - oldPassword
                - newPassword
              properties:
                oldPassword:
                  type: string
                  format: password
                  description: Current password
                  example: currentPassword123
                newPassword:
                  type: string
                  format: password
                  minLength: 8
                  description: New password (minimum 8 characters)
                  example: newPassword456
      responses:
        '200':
          description: Password updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                message: Password updated successfully
        '400':
          description: Missing fields or password too short
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                missingFields:
                  value:
                    success: false
                    message: Old password and new password are required
                tooShort:
                  value:
                    success: false
                    message: New password must be at least 8 characters long
        '401':
          description: Current password is incorrect or unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                incorrectPassword:
                  value:
                    success: false
                    message: Current password is incorrect
                unauthorized:
                  value:
                    success: false
                    message: Invalid or expired token
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: User not found
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: Failed to update password
  /profile/edit/email:
    put:
      tags:
        - Profile
      summary: Change email address
      description: Change the authenticated user's email address
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - oldEmail
                - newEmail
              properties:
                oldEmail:
                  type: string
                  format: email
                  description: Current email address
                  example: current@example.com
                newEmail:
                  type: string
                  format: email
                  description: New email address
                  example: new@example.com
      responses:
        '200':
          description: Email updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                message: Email updated successfully
        '400':
          description: Missing fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: Old email and new email are required
        '401':
          description: Current email is incorrect or unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                incorrectEmail:
                  value:
                    success: false
                    message: Current email is incorrect
                unauthorized:
                  value:
                    success: false
                    message: Invalid or expired token
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: User not found
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: Failed to update email
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or registration
  schemas:
...
    SuccessResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation successful
    Error:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: An error occurred
    ValidationError:
      type: object
      required:
        - success
        - message
        - errors
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Validation failed
        errors:
          type: array
          items:
            type: string
          example: ["Email is required", "Password must be at least 8 characters"]
    InvalidCategoriesError:
      type: object
      required:
        - success
        - message
        - invalidCategories
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Invalid categories provided
        invalidCategories:
          type: array
          items:
            type: string
          example: ["invalid_category_1", "invalid_category_2"]

    # User schemas
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
          example: 1
        userName:
          type: string
          nullable: true
          example: JohnDoe
        currentLocation:
          type: string
          nullable: true
          example: New York, NY
        bio:
          type: string
          nullable: true
          example: Software developer and basketball enthusiast
        createdAt:
          type: string
          format: date-time
          example: '2024-01-15T10:30:00.000Z'
        updatedAt:
          type: string
          format: date-time
          example: '2024-02-11T10:30:00.000Z'
        lastLogin:
          type: string
          format: date-time
          nullable: true
          example: '2024-02-12T08:30:00.000Z'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
    Category:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: basketball

    # Event schemas
...
    ProfileSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            user:
              $ref: '#/components/schemas/UserProfile'
    ProfileUpdateSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            user:
              type: object
              properties:
                id:
                  type: integer
                  example: 1
                email:
                  type: string
                  format: email
                  example: user@example.com
                userName:
                  type: string
                  example: JohnDoe_Updated
                currentLocation:
                  type: string
                  nullable: true
                  example: Los Angeles, CA
                bio:
                  type: string
                  nullable: true
                  example: Full-stack developer
                categories:
                  type: array
                  items:
                    type: string
                  example: ["basketball", "football", "programming"]
                updatedAt:
                  type: string
                  format: date-time
                  example: '2024-02-12T10:30:00.000Z'
...